docs: seperate popup content
